refactor: replace useSampleSongs hook with useSampleChordSheets; update related components
